+++ app/controllers/api/v1/shop/orders_controller.rb
@@ -1,16 +1,18 @@
 class Api::V1::Shop::OrdersController < ApiController
   before_action :authenticate_user!
   before_action :set_delivery, only: [:choose_delivery]
-  before_action :set_order, only: [:show, :update, :destroy]
+  before_action :set_order, only: [:show, :update, :destroy, :delivery, :choose_delivery]
   after_action :verify_authorized
 
   def index
     @orders = Order.where(shop_id: params[:shop_id])
+    authorize @orders.first
     render json: Api::V1::Shop::OrderSerializer.new(@orders).serialized_json
   end
 
   def create
     @order = Order.new order_params
+    authorize @order
     if @order.save
       render json: {success: true,message: 'Order Successfully Created!'},status: 200
     else
@@ -19,6 +21,7 @@ class Api::V1::Shop::OrdersController < ApiController
   end
 
   def show
+    authorize @order
     if @order
       render json: Api::V1::Shop::OrderSerializer.new(@order).serialized_json
     else
@@ -27,6 +30,7 @@ class Api::V1::Shop::OrdersController < ApiController
   end
 
   def update
+    authorize @order
     if @order.update order_params
       render json: Api::V1::Shop::OrderSerializer.new(@order).serialized_json
     else
@@ -35,11 +39,13 @@ class Api::V1::Shop::OrdersController < ApiController
   end
 
   def delivery
+    authorize @order
     @deliveries = User.where(user_type: 'delivery')
     render json: Api::V1::UserSerializer.new(@deliveries).serialized_json
   end
 
   def choose_delivery
+    authorize @order
     if @delivery.update delivery_params
       render json: {success: true,message: 'Delivery Successfully Selected!'},status: 200
     else
@@ -48,6 +54,7 @@ class Api::V1::Shop::OrdersController < ApiController
   end
 
   def destroy
+    authorize @order
     if @order.destroy
       render json: {success: true,message: 'Collection Successfully Deleted!'},status: 200
     else
